Attempt to save state on quit and merge updates
When starting the application, try to load
$USER_CACHE_DIR/todoist-tui/state.json and set that as the state. Using
the sync token in state.json, make a full sync with that sync token to
get an incremental update. Merge in the new state with the state in
state.json. Only merges in the tasks right now, but should eventually
merge in everything once I understand the semantics of stuff better

Signed-off-by: Tom George <tg82490@gmail.com>